Get everything else moved around properly
