+++ packages/desktop/app/components/PatientPrinting/modals/PatientPrintDetailsModal.js
@@ -7,8 +7,8 @@ import { Colors } from '../../../constants';
 import { useApi, isErrorUnknownAllow404s } from '../../../api';
 import { useLocalisation } from '../../../contexts/Localisation';
 
-import { PatientIDCardPage } from '../reusable/PatientIDCardPage';
-import { PatientStickerLabelPage } from '../reusable/PatientStickerLabelPage';
+import { PatientIDCardPage } from './PatientIDCardPage';
+import { PatientStickerLabelPage } from './PatientStickerLabelPage';
 import { CovidTestCertificateModal } from './CovidTestCertificateModal';
 import { CovidClearanceCertificateModal } from './CovidClearanceCertificateModal';
 import { BirthNotificationCertificateModal } from './BirthNotificationCertificateModal';
