+++ packages/desktop/app/components/PatientPrinting/modals/PatientStickerLabelPage.js
@@ -1,13 +1,13 @@
 import React, { useEffect } from 'react';
 import styled from 'styled-components';
 
-import { SEX_VALUE_INDEX } from '../../../../constants';
-import { useElectron } from '../../../../contexts/Electron';
-import { DateDisplay } from '../../../DateDisplay';
-import { useLocalisation } from '../../../../contexts/Localisation';
+import { SEX_VALUE_INDEX } from '../../../constants';
+import { useElectron } from '../../../contexts/Electron';
+import { DateDisplay } from '../../DateDisplay';
+import { useLocalisation } from '../../../contexts/Localisation';
 
-import { PrintPortal } from './PrintPortal';
-import { PatientBarcode } from './PatientBarcode';
+import { PrintPortal } from '../PrintPortal';
+import { PatientBarcode } from '../printouts/reusable/PatientBarcode';
 
 const Sticker = styled.div`
   font-family: monospace;
